rename proxies context
